2019 Offseason Development Update 1
This update is the result of weeks' worth of effort. As a result of  changes (some major, some minor) being made here and there in several files, our robot's capabilities have expanded. For example, it can now perform multiple tasks autonomously using commands. I recommend checking them out (they're in the "commands" folder) because they're the best examples of working commands we have as of today. To test them out, you can link commands to certain buttons on a controller using the [button].get() method to check for controller input. Alternatively, you can try to activate them on the SmartDashboard (you should see several command options, which are added using m_chooser.addOption) by clicking the command buttons on there while connected to the robot. You might notice that the SmartDashboard is a lot less cluttered, as some unnecessary/useless items were removed. As for the subsystems, upon which the commands rely, a couple of them had their functions modified in order to better fit our robot's capabilities. All in all, that's basically the extent of the changes that were made. Check the list of updated files below.

Commands / Command Groups
-ArmToUpLimit.java brings our robot's arm up
-AutonCrossHabLine.java drives our robot forward without turning
-CommandGroupTest.java tests multiple commands in sequence
-hatchIn.java brings in the hatch panel manipulator
-hatchOut.java brings out the hatch panel manipulator
-pistonIn.java aims to retract the piston
-pistonOut.java aims to extend the piston

Subsystems
-DriveTrain.java is our robot's drivetrain subsystem that now uses arcade drive
-exampleSolenoid.java is a subsystem that covers our robot's solenoid functions

Robot
-Robot.java saw a lot of changes regarding the SmartDashboard, like commands being added for selection when using the robot.