Drivetrain Revamp + other subsystems
Our drivetrain has seen the most change out of all our programs this round of updates.  To start, our drivetrain is now more versatile as we can use either tank or arcade drive to move our robot as well as stop the motors from within the subsystem. Furthermore, master and follower motors are outlined within, and we can now create drivetrain objects with this subsystem.  However, most of what was added comes from code that I worked on as a member of Team 3414, and I recommend that you check out the program yourself because there are almost too many new things to describe. I'll put it simply. First, the turnRadius() method moves the robot in a circle with variable radius and rotation.  Next, the moveStraight() method drives the robot straight and will correct the robot's alignment should it be a bit crooked.  Third, the configSpeedPlan() method creates speedplans for methods that move the robot to follow. Lastly, our moveGyro() method essentially moves the robot a given distance while rotating it; speed and direction should be fed into the function as well. PLEASE read the code thoroughly to gain a complete understanding of it.