Robot.java
In the Robot.java file, three types of changes have been made.  First, with the addition of encoders to our robot, code regarding the sensors and their feedback was put in. In addition, our robot program now calls the getInstance() methods that some of our commands have for more efficient command calling. Lastly, the scheduler, which organizes and calls commands (options) that you want to run autonomously, option declarations were moved further down in the robotInit() method for sequencing purposes.