MoveGyro.java
You might have noticed that this program is also in the Drivetrain.java file. Because MoveGyro allows the robot to move varying distances in practically any direction we want it to and at variable speeds, I thought it to be important enough to warrant its own command. Feedback from encoders (distance) and the navX (yaw) is constantly read to help guide the robot to travel the entire distance we tell it to do so, and at the speed, angle, and direction we want. There are many ways we can use MoveGyro, so expect to see it being used a lot later.